Failing tests for simple ContactImpl constructor
The assignment description dictates that we have two constructors; this
one only takes `id` and `name` parameters.

I've included a 'happy path' test case, which isn't necessary for the
more complex constructor because it's already used in the `setUp()`
method.

I'd like to tidy these tests later as the structure isn't ideal as-is.