Split past/future meetings into separate variables
Split past and future meetings into separate data structures to avoid
problems with type conversions.

I wanted to avoid doing this since time is continuously changing and the
difference between a 'past' meeting and a 'future' meeting depends on
the user's system time. I'm hoping to come back to this later.