Make inner class static
This inner class doesn't reference the object it's embedded in, so we
can make it static, which would reduce memory usage if TestMeetingImpl
were instantiated more than once.

This was highlighted by FindBugs. It's not really necessary, but it does
mean FindBugs will complain less when the teaching staff are marking
this project :-)